0.17.2-ALPHA - optimized requests to localization bundle - added new class BundleFillerExtension - improvements of BrowserProxyCriteria
