+++ build.gradle
@@ -1,5 +1,5 @@
 ext {
-    globalVersion = '0.17.1-ALPHA'
+    globalVersion = '0.17.2-ALPHA'
 }
 
 subprojects {
