+++ core.api/doc/rus/LOCALIZATION.MD
@@ -80,36 +80,31 @@ core.api.steps.SequentialGetStepSupplier.resultDescription = Result
 ######################## core.api.steps.SequentialActionSupplier #
 #_________________________________Parameters_____________________________________
 #Original text = Perform:
-core.api.steps.SequentialActionSupplier.imperative = Perform:
-
+core.api.steps.SequentialActionSupplier.imperative=Perform:
 ######################## core.api.steps.Absence #
 #_________________________________Parameters_____________________________________
 #Original text = Time of the waiting for absence
-core.api.steps.Absence.timeOut = Time of the waiting for absence
+core.api.steps.Absence.timeOut=Time of the waiting for absence
 #Original text = Is absent
-core.api.steps.Absence.resultDescription = Is absent
+core.api.steps.Absence.resultDescription=Is absent
 #__________________________________ Methods _______________________________________
 #Original text = Absence of {toBeAbsent}
-core.api.steps.Absence.absence(core.api.steps.SequentialGetStepSupplier<T,?,?,?,?>) = Absence of {toBeAbsent}
+core.api.steps.Absence.absence(java.util.function.Function<T,?>)=Absence of {toBeAbsent}
 #Original text = Absence of {toBeAbsent}
-core.api.steps.Absence.absence(java.util.function.Function<T,?>) = Absence of {toBeAbsent}
-
+core.api.steps.Absence.absence(core.api.steps.SequentialGetStepSupplier<T,?,?,?,?>)=Absence of {toBeAbsent}
 ######################## core.api.steps.Presence #
 #_________________________________Parameters_____________________________________
 #Original text = Is present
-core.api.steps.Presence.resultDescription = Is present
+core.api.steps.Presence.resultDescription=Is present
 #__________________________________ Methods _______________________________________
 #Original text = Presence of {toBePresent}
-core.api.steps.Presence.presence(java.util.function.Function<T,?>) = Presence of {toBePresent}
+core.api.steps.Presence.presence(core.api.steps.SequentialGetStepSupplier<T,?,?,?,?>)=Presence of {toBePresent}
 #Original text = Presence of {toBePresent}
-core.api.steps.Presence.presence(core.api.steps.SequentialGetStepSupplier<T,?,?,?,?>) = Presence of {toBePresent}
-
-#======================================= ATTACHMENTS ============================================
-
+core.api.steps.Presence.presence(java.util.function.Function<T,?>)=Presence of {toBePresent}
+#============================================ ATTACHMENTS ============================================ 
 ######################## core.api.event.firing.collections.ArrayCaptor #
 #Original text = Resulted array
-core.api.event.firing.collections.ArrayCaptor = Resulted array
-
+core.api.event.firing.collections.ArrayCaptor=Resulted array
 ######################## core.api.event.firing.collections.CollectionCaptor #
 #Original text = Resulted collection
 core.api.event.firing.collections.CollectionCaptor = Resulted collection
@@ -117,7 +112,7 @@ core.api.event.firing.collections.CollectionCaptor = Resulted collection
 ######################## core.api.event.firing.collections.MapCaptor #
 #Original text = Resulted map
 core.api.event.firing.collections.MapCaptor = Resulted map
-  ```
+```
 
 ### Использовать aspectj для собственных расширений механизма шагов
