+++ selenium/src/main/java/ru/tinkoff/qa/neptune/selenium/functions/searching/SearchSupplier.java
@@ -33,7 +33,7 @@ import static ru.tinkoff.qa.neptune.selenium.properties.WaitingProperties.ELEMEN
 @SequentialGetStepSupplier.DefineCriteriaParameterName("Element criteria")
 @SequentialGetStepSupplier.DefineGetImperativeParameterName(value = "Find:")
 @SequentialGetStepSupplier.DefineResultDescriptionParameterName("Found element")
-@MaxDepthOfReporting(1)
+@MaxDepthOfReporting(0)
 @CaptureOnSuccess(by = WebElementImageCaptor.class)
 @CaptureOnFailure(by = {WebDriverImageCaptor.class, WebElementImageCaptor.class})
 public final class SearchSupplier<R extends SearchContext>
